+++ src/TemplateManager.cs
@@ -29,6 +29,7 @@ using static Il2CppMono.Security.X509.X520;
 using Harmony;
 using Il2CppScheduleOne.ItemFramework;
 using Il2CppScheduleOne.Persistence;
+using DeliverySaver.src;
 
 namespace DeliverySaver
 {
@@ -73,6 +74,7 @@ namespace DeliverySaver
     {
         private Dictionary<string, List<EntryData>> _toSaves = new Dictionary<string, List<EntryData>>();
         private Template _template;
+        private RegisteredEntry _registeredEntry;
         private static TemplateManager _instance;
         public static TemplateManager Instance
         {
@@ -93,42 +95,88 @@ namespace DeliverySaver
             AssetsManager.Instance.LoadAssetBundleFromResources("Template", "ui.template");
             AssetsManager.Instance.LoadAssetBundleFromResources("Entry", "ui.entry");
             AssetsManager.Instance.LoadAssetBundleFromResources("Component", "ui.component");
+            AssetsManager.Instance.LoadAssetBundleFromResources("RegisteredEntryPrompt", "ui.registeredentryprompt");
+            AssetsManager.Instance.LoadAssetBundleFromResources("RegisteredAllEntryPrompt", "ui.registeredallentryprompt");
         }
 
         public void CreateTemplateGameObject()
         {
             _template = new Template();
+            _registeredEntry = new RegisteredEntry("RegisteredEntryPrompt", "RegisteredAllEntryPrompt");
         }
 
-        public void AddEntryData(EntryData entry, bool showSameEntryMessage = true)
+        public bool AddEntryData(EntryData entry)
         {
             var entryData = FindDataFromTitle(entry.title);
 
             if (entryData != null)
             {
-                if (showSameEntryMessage)
-                {
-                    throw new EntryAlreadyExistsException();
-                }
-
-                return;
+                _registeredEntry.Show(entry);
+                return false;
             }
 
             AddEntryDataBase(entry);
             _toSaves[GameInfo.Instance.GameName].Add(entry);
+            return true;
         }
 
         private void AddEntryDataBase(EntryData entry)
         {
             _template.AddEntryData(entry);
             _template.RebuildLayout();
+            _template.Open();
         }
 
-        public void AddEntry(string name, DeliveryShop shop)
+        public void AddEntriesData(List<EntryData> entries)
         {
+            var sames = entries.Intersect(_toSaves[GameInfo.Instance.GameName]);
+            var uniques = entries.Except(sames).ToList();
+
+            foreach (var entry in uniques)
+            {
+                AddEntryData(entry);
+            }
+
+            _registeredEntry.Show(sames);
+        }
+
+        public bool AddEntry(string name, DeliveryShop shop)
+        {
+            var entryData = FindDataFromTitle(name);
+
+            if (entryData != null)
+            {
+                _registeredEntry.Show(EntryData.FromDeliveryShop(name, entryData.multiplier, shop));
+                return false;
+            }
+
             _template.AddEntry(name, shop);
             _toSaves[GameInfo.Instance.GameName].Add(_template.GetLastEntry().ToEntryData());
             _template.RebuildLayout();
+            _template.Open();
+
+            return true;
+        }
+
+        public void SetEntry(Entry newEntry)
+        {
+            EntryData newEntryData = newEntry.ToEntryData();
+            SetEntryData(newEntryData);
+        }
+
+        public void SetEntryData(EntryData newEntryData)
+        {
+            int index = _toSaves[GameInfo.Instance.GameName].FindIndex(ed => ed.title == newEntryData.title);
+
+            if (index != -1)
+            {
+                _toSaves[GameInfo.Instance.GameName][index] = newEntryData;
+                _template.SetEntryData(newEntryData);
+                _template.RebuildLayout();
+                return;
+            }
+
+            throw new Exception($"EntryData {newEntryData.title} has not been found");
         }
 
         public void UpdateEntryTitle(string oldTitle, Entry entry)
@@ -200,6 +248,8 @@ namespace DeliverySaver
                     AddEntryData(entry);
                 }
             }
+
+            _template.Close();
         }
 
         internal void RemoveEntry(Entry entry)
@@ -218,7 +268,7 @@ namespace DeliverySaver
             return _toSaves[GameInfo.Instance.GameName].Find(e => e.title == title);
         }
 
-        internal bool IsEntryRegister(string title)
+        public bool IsEntryRegister(string title)
         {
             return _toSaves[GameInfo.Instance.GameName].Find(e => e.title == title) != null;
         }